Completed section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -13,6 +13,7 @@ using Microsoft.AspNetCore.Http.HttpResults;
 using AutoMapper.QueryableExtensions;
 using MassTransit;
 using Contracts;
+using Microsoft.AspNetCore.Authorization;
 
 namespace AuctionService.Controllers
 {
@@ -62,12 +63,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
-            // TODO: Add user as seller
-            auction.Seller = "test";
+
+            auction.Seller = User.Identity.Name;
 
             _context.Add(auction);
 
@@ -77,14 +79,13 @@ namespace AuctionService.Controllers
 
             var result = await _context.SaveChangesAsync() > 0;
 
-            
-
             if (!result)
                 return BadRequest("Could not save changes to the DB");
 
             return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -95,7 +96,8 @@ namespace AuctionService.Controllers
             if (auction == null)
                 return NotFound();
 
-            // TODO: check seller == username
+            if (auction.Seller != User.Identity.Name)
+                return Forbid();
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -113,6 +115,7 @@ namespace AuctionService.Controllers
             return BadRequest("Problem saving changes.");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -121,7 +124,8 @@ namespace AuctionService.Controllers
             if (auction == null)
                 return NotFound();
 
-            // TODO: ckeck seller == username
+            if (auction.Seller != User.Identity.Name)
+                return Forbid();
 
             _context.Auctions.Remove(auction);
